requirements.txt
antiorm==1.2.1
blinker==1.7.0
click==8.1.7
Flask==3.0.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
pyyaml==6.0.1
